+++ jupyter_ai_agents/providers/azure_openai.py
@@ -2,26 +2,33 @@
 #
 # BSD 3-Clause License
 
-from langchain_openai import AzureChatOpenAI
 from langchain.agents import AgentExecutor, create_openai_tools_agent
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
+from langchain_openai import AzureChatOpenAI
 
 
-def create_azure_openai_agent(model_name: str, system_prompt: str, tools: list) -> dict:
+def create_azure_openai_agent(model_name: str, system_prompt: str, tools: list) -> AgentExecutor:
     """Create an agent from a set of tools and an Azure deployment"""
-    
+
     llm = AzureChatOpenAI(azure_deployment=model_name)
 
     prompt = ChatPromptTemplate.from_messages(
-    [
-        ("system", system_prompt),
-        ("user", "{input}"),
-        MessagesPlaceholder(variable_name="agent_scratchpad"),
-    ])
+        [
+            ("system", system_prompt),
+            ("user", "{input}"),
+            MessagesPlaceholder(variable_name="agent_scratchpad"),
+        ]
+    )
 
     # Create the agent using LangChain's built-in tool handling
     agent = create_openai_tools_agent(llm, tools, prompt)
-    
-    agent_executor = AgentExecutor(name="NotebookPromptAgent", agent=agent, tools=tools, verbose=True, handle_parsing_errors=True)
+
+    agent_executor = AgentExecutor(
+        name="NotebookPromptAgent",
+        agent=agent,
+        tools=tools,
+        verbose=True,
+        handle_parsing_errors=True,
+    )
 
     return agent_executor
